+++ webapp/src/app/components/professor/course-details/course-details.component.ts
@@ -146,29 +146,35 @@ export class CourseProfessorDetailsComponent {
   }
 
   deleteSurveyFromCourse(survey: Survey) {
-    this.apiProfessorService
-      .deleteSurveyFromCourse(this.course.id, survey.id!)
-      .subscribe((msg: Message) => {
-        alert(msg.message)
-        this.surveys = this.surveys.filter((s) => s.id != survey.id)
-      })
+    if (confirm("Seguro que quieres eliminar el cuestionario del curso")) {
+      this.apiProfessorService
+        .deleteSurveyFromCourse(this.course.id, survey.id!)
+        .subscribe((msg: Message) => {
+          alert(msg.message)
+          this.surveys = this.surveys.filter((s) => s.id != survey.id)
+        })
+    }
   }
 
   deleteQuestionFromCourse(question: Question) {
-    this.apiProfessorService
-      .deleteQuestionFromCourse(this.course.id, question.id)
-      .subscribe((msg: Message) => {
-        alert(msg.message)
-        this.questions = this.questions.filter((q) => q.id != question.id)
-      })
+    if (confirm("Seguro que quieres eliminar la pregunta del curso")) {
+      this.apiProfessorService
+        .deleteQuestionFromCourse(this.course.id, question.id)
+        .subscribe((msg: Message) => {
+          alert(msg.message)
+          this.questions = this.questions.filter((q) => q.id != question.id)
+        })
+    }
   }
 
   deleteUserFromCourse(user: User) {
-    this.apiProfessorService
-      .deleteUserFromCourse(this.course.id, user.id)
-      .subscribe((msg: Message) => {
-        alert(msg.message)
-        this.users = this.users.filter((u) => u.id != user.id)
-      })
+    if (confirm("Seguro que quieres eliminar el usuario del curso")) {
+      this.apiProfessorService
+        .deleteUserFromCourse(this.course.id, user.id)
+        .subscribe((msg: Message) => {
+          alert(msg.message)
+          this.users = this.users.filter((u) => u.id != user.id)
+        })
+    }
   }
 }
